V2.0.0 (#36)
* Issue #27: OpenAI API access; Centralized route still remaining.

Added a openai API method to call completion service. Also added the necessary settings required to call this API in the example.config.json file.

Second part where we add a route so that it can be called from other routes will follow next.

* Issue #25: Expandible task card

In this contribution I have made the task card expandible so that description can be read in the home page itself rather than going to view task route.

* Bug #22: Fixing date card for old dates

Date cards used to show even for older dates with negative days countdown. This is now fixed by comparing the date with today's date, if the date is older by 1 day then the date card is not rendered.

* #34: Added enhancement to show warning if tasks are due in less than or equal to X days

* [Daily Log]: Tasks that were closed within OPENED_TASK_LOG_DAYS appear in log form.

* [Daily Log]: Tasks that were closed within OPENED_TASK_LOG_DAYS appear in log form.

* #24: Make tags click searchable

---------

Co-authored-by: Amrit Pandey <amrit.pandey@sap.com>